+++ android/app/src/main/java/io/github/beingmechon/babytracker/MainActivity.java
@@ -1,0 +1,5 @@
+package io.github.beingmechon.babytracker;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}